Add support for SailJail
WARNING: this is just an implementation idea.  Use with caution!

You need to copy (or better yet, hardlink) your keys to the new
location manually or risk data loss.
I have successfully tested this and verified that it works (FOR MY USE).

Note that in order to use it, you need to disable SailJail detection
that happens in this app, too.